Chatbot-Javeed/requirements.txt
faiss-cpu
groq 
langchain-groq 
PyPDF2
langchain_google_genai 
langchain 
langchain_community 
streamlit 
python-dotenv
pypdf

langchain-text-splitters
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
langchain_community
streamlit
python-dotenv
pypdf
langchain-text-splitters
flask
flask_cors
langchain-mistralai
gunicorn
